Continue Loop
moving the first variable to the next value in the sequence and the second variable to the sum. This will continue while t4 is less than or equal to the uses entered value (t3). when t4 is over that amount t5 will be set to zero and the loop will end.